requirements.txt
adal==1.2.6
anyio==2.1.0
anytree==2.8.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
audioread==2.1.9
azure-common==1.1.26
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.0.0
azure-mgmt-storage==11.2.0
azureml-core==1.23.0
Babel==2.9.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker==4.4.4
entrypoints==0.3
experta==1.9.4
freetype-py==2.2.0
frozendict==1.2
glfw==2.0.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.4.0
jupyterlab==3.0.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
msrest==0.6.21
msrestazure==0.6.4
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
ndg-httpsclient==0.5.1
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
p5==0.7.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
pyasn1==0.4.8
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
PyJWT==2.0.1
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==227
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.6.7
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
SecretStorage==3.3.1
Send2Trash==1.5.0
SimpleGUICS2Pygame==2.1.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
vispy==0.6.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
